Fix player active on reset game
Fix issue where the 'Player active' class is incorrectly added to the wrong player when starting a new game. The problem was caused by the toggle class function, which was toggling the class on the wrong player. To resolve this, I explicitly targeted the specific player during the game reset, ensuring that the 'Player active' class is correctly assigned
Fix player active on reset game